Mod: バッチ処理API用のエラーハンドラーを追加し、全タグ一括登録APIのエラーハンドリングを統一
